+++ CHANGELOG.md
@@ -23,7 +23,9 @@
 
 #### LSP
 
-- add ``commandArgument`` semantic token type for better semantic highlight of command argument
+- improve ``textDocument/semanticTokens``
+    - add ``commandArgument`` semantic token type for better semantic highlight of command argument
+    - check client capability and now only emit client supported token types
 
 #### API